+++ crates/solar_sonar/src/cli.rs
@@ -9,23 +9,41 @@ pub struct Cli {
     /// Jump range of systems
     #[clap(default_value_t = Args::DEFAULT_JUMPS)]
     pub jumps: u8,
-    #[clap(default_value_t = Args::DEFAULT_STDIO)]
+    /// Enable / disable output to terminal
+    #[clap(long, default_value_t = Args::DEFAULT_STDIO, action = clap::ArgAction::Set)]
     pub stdio: bool,
+    // Enable / disable audio alerts
+    #[clap(long, default_value_t = Args::DEFAULT_AUDIO, action = clap::ArgAction::Set)]
+    pub audio: bool,
+    /// Replay a specific log file
+    #[clap(long)]
+    pub replay: Option<PathBuf>,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct Args {
     pub watch_character_ids: Vec<CharacterID>,
     pub watch_system_ids: Vec<SolarID>,
     pub jumps: u8,
+    pub audio: bool,
     pub stdio: bool,
+    pub replay_file: Option<PathBuf>,
+}
+
+#[derive(Debug)]
+pub struct ArgParam {
+    pub args: Args,
+    pub chat_channels: Option<ChatChannels>,
 }
 
 impl Args {
     pub const DEFAULT_JUMPS: u8 = 2;
+    pub const DEFAULT_AUDIO: bool = true;
     pub const DEFAULT_STDIO: bool = true;
+}
 
-    pub(crate) fn try_from_cli(cli: Cli, cfg: &Config) -> SolarResult<Self> {
+impl Args {
+    pub(crate) fn try_from_cli(cli: Cli, cfg: &Config) -> SolarResult<ArgParam> {
         let jumps = cli.jumps;
         let watch_character_ids = cli.watch_characters.split(',')
             .map(|snake| snake.to_snake_case())
@@ -44,14 +62,35 @@ impl Args {
             )
             .collect::<SolarResult<Vec<_>>>()?;
 
-        Ok(Self {
+        let replay_file = cli.replay
+            .and_then(|p| Some(expand_pathbuf(p)))
+            .transpose()?;
+
+        let this = Args {
             watch_character_ids,
             watch_system_ids,
             jumps,
-            stdio: true,
-        })
+            audio: cli.audio,
+            stdio: cli.stdio,
+            replay_file,
+        };
+
+        this.build()
     }
 
+    pub fn build(self) -> SolarResult<ArgParam> {
+        let chat_channels = self.replay_file.as_ref().map(PathBuf::from)
+            .and_then(|f| ChatLogFile::from_path_buf(f))
+            .map(|f| ChatChannels::new(vec![f.channel().to_string()]));
+
+        Ok(ArgParam {
+            args: self,
+            chat_channels,
+        })
+    }
+}
+
+impl Args {
     pub(crate) fn watch_characters<'a>(&self, cfg: &'a Config) -> Vec<&'a CharacterConfig> {
         self.watch_character_ids.iter()
             .filter_map(|id| cfg.characters.iter().find(|chr| &chr.id == id))
@@ -62,6 +101,10 @@ impl Args {
         self.watch_system_ids.iter()
             .map(|id| STAR_MAP.system(id))
             .collect()
+    }
+
+    pub(crate) fn is_replay(&self) -> bool {
+        self.replay_file.is_some()
     }
 }