+++ docs/CHANGES.md
@@ -1,4 +1,9 @@
 # CHANGES: solar_sonar
 
+## 0.2.0
+- New command-line features:
+  - Perform `--replay` of a specific log file
+  - Disable output of `--stdio` and `--audio`
+- Basic library support via `solar_sonar::start`
 ## 0.1.1
 First release.